feat: add ImageIndexIVFPQ and group indexes under retrieval.index
Add a product-quantization IVF index wrapping faiss.IndexIVFPQ. On top of the
shared nlist/nprobe it accepts m (number of PQ sub-vectors) and nbits (bits
per sub-vector), validating that m divides the feature dimensionality and that
the gallery has at least 2**nbits vectors to train the PQ codebook. It exposes
the coarse-quantizer centroids as cluster_centers, like IVFFlat.

Move the index modules into a retrieval/index subpackage so they import as
``from pyvisim.retrieval.index import ...``; pyvisim.retrieval re-exports them
for convenience.
