+++ pyvisim/retrieval/index/__init__.py
@@ -1,0 +1,6 @@
+"""FAISS-backed image indexes for accelerated similarity search."""
+
+from ._base_index import ImageIndex
+from .image_index import ImageIndexIVFFlat, ImageIndexIVFPQ
+
+__all__ = ["ImageIndex", "ImageIndexIVFFlat", "ImageIndexIVFPQ"]
